refactor(cli): extract config module
